+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2020-10-22
+### Fixed
+- Algorithm specific configuration load failure now throws an error instead of silently creation of empty config.
+- The _fit_ command query with an empty target column provides `None` to a model implementation instead of `Some(null)`.
+- The service columns (it is starts and ends with double underscore character) drop from the resulting data frame of the _apply_'s command.
+### Added
+- Test for the _fit_ command.
+- Test for the _apply_ command.
+- Test for the _score_ command.
+
 ## [2.1.0] - 2020-10-16
 ### Changed
 - Algorithms now is separated from this SMaLL core plugin.